Update zabbix_alerta.py
Seems these 2 attributes need to be set for the alerta_zabbix.py plugin on the Alerta end of this to be able to Ack properly.   This was missing from the examples here and in the documentation.

Otherwise the alerta plugin will return when it is not set as per this code from the alerta_zabbix plugin. 
        event_id = alert.attributes.get('eventId', None)
        trigger_id = alert.attributes.get('triggerId', None)
        more_info = alert.attributes.get('moreInfo',None)

        if not event_id:
            return